+++ R/p_value_kenward.R
@@ -21,10 +21,9 @@
 #' @seealso `dof_kenward()` and `se_kenward()` are small helper-functions
 #' to calculate approximated degrees of freedom and standard errors for model
 #' parameters, based on the Kenward-Roger (1997) approach.
-#' \cr \cr
-#' [`dof_satterthwaite()`][dof_satterthwaite] and
-#' [`dof_ml1()`][dof_ml1] approximate degrees
-#' of freedom based on Satterthwaite's method or the "m-l-1" rule.
+#'
+#' [`dof_satterthwaite()`] and [`dof_ml1()`] approximate degrees of freedom
+#' based on Satterthwaite's method or the "m-l-1" rule.
 #'
 #' @examples
 #' \donttest{